Move change log into a partial
This allows the partial to be only the raw, Markdown source of the
change log. It is an exact copy of the source from the Bourbon repo.

We can pull that into a page and use the page to wrap it will HTML and
classes for styling.

This should make it easier to automate the updating of this file when
generating new doc versions.

See: https://github.com/thoughtbot/bourbon.io/issues/14