Add handling for undefined group names in legacy docs
Because they predate the use of sassdoc for documentation, older versions of
neat (`~>'1.8'`) do not have groups defined for the various features. This
caused the group names to read as "Undefined". This change checks to see if
they are undefined and if it is, the group heading will be omitted.

https://github.com/thoughtbot/neat/issues/564